+++ solutions/src/_2024/_example/mod.rs
@@ -6,11 +6,11 @@ pub fn run() -> super::Runner {
 }
 
 // Answer: ???
-fn part_1() -> u32 {
+fn part_1() -> u64 {
     0
 }
 
 // Answer: ???
-fn part_2() -> u32 {
+fn part_2() -> u64 {
     0
 }
